requirements.txt
pyserial>=3.5
openai>=1.0.0
requests>=2.28.0
